travel-planner-a2a protocol/pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-ollama>=0.3.6",
    "python-a2a>=0.5.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "tavily-python>=0.7.10",
]
